Refactor: Move Environment to `lib/`
