+++ Resources/Private/JavaScript/lib/Environment.js
